requirements.txt
crewai[tools]>=0.100.0
python-dotenv>=1.0.0
rich>=13.0.0
